+++ localflavor/ir/ir_provinces.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.utils.translation import ugettext_lazy as _
+
+#: A list of Iran provinces and autonomous cities as `choices` in a
+#: formfield. From https://en.wikipedia.org/wiki/Provinces_of_Iran
+PROVINCE_CHOICES = (
+    ('AL', _('Alborz')),
+    ('AR', _('Ardabil')),
+    ('AE', _('Azerbaijan East')),
+    ('AW', _('Azerbaijan Wast')),
+    ('BU', _('Bushehr')),
+    ('CM', _('Chahar Mahaal and Bakhtiari')),
+    ('FA', _('Fars')),
+    ('GI', _('Gilan')),
+    ('GO', _('Golestan')),
+    ('HA', _('Hamadan')),
+    ('HO', _('Hormozgan')),
+    ('IL', _('Ilam')),
+    ('IS', _('Isfahan')),
+    ('KE', _('Kerman')),
+    ('KM', _('Kermanshah')),
+    ('KN', _('Khorasan North')),
+    ('KR', _('Khorasan Razavi')),
+    ('KS', _('Khorasan South')),
+    ('KH', _('Khuzestan')),
+    ('KB', _('Kohgiluyeh and Boyer-Ahmad')),
+    ('KU', _('Kurdistan')),
+    ('LO', _('Lorestan')),
+    ('MA', _('Markazi')),
+    ('MZ', _('Mazandaran')),
+    ('QA', _('Qazvin')),
+    ('QO', _('Qom')),
+    ('SE', _('Semnan')),
+    ('SB', _('Sistan and Baluchestan')),
+    ('TH', _('Tehran')),
+    ('YZ', _('Yazd')),
+    ('ZN', _('Zanjan')),
+)